requirements.txt
pytesseract
Pillow
pdf2image
ai.engine==0.0.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
apispec==6.6.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bech32==1.2.0
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cosmpy==0.9.2
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
-e .
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.9.0
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-groq==0.1.6
langchain-text-splitters==0.2.2
langsmith==0.1.93
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.6
packaging==24.1
protobuf==4.25.3
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
ratelim==0.1.6
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.37.2
tenacity==8.5.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uagents==0.14.0
uagents-ai-engine==0.4.0
urllib3==2.2.2
uvicorn==0.30.3
watchfiles==0.22.0
websockets==10.4
yarl==1.9.4
